pyproject.toml
[project]
name = "ai-tutor"
version = "0.1.0"
description = "A Streamlit AI Tutor app"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "AI Tutor Team" }
]

dependencies = [
  "streamlit>=1.33",
  "openai>=1.30.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]


